refactor: avoid same name mode with crate
