+++ manifest_reader/src/manifest/node/definition.rs
