+++ manifest_reader/src/manifest/node/mod.rs
@@ -1,2 +1,2 @@
+pub mod definition;
 pub mod input_from;
-pub mod node;
