+++ manifest_reader/src/path_finder/finder.rs
