+++ manifest_reader/src/path_finder/mod.rs
@@ -1,12 +1,12 @@
 mod block;
+mod finder;
 mod flow;
 mod manifest_file;
 mod package;
-mod path_finder;
 mod search_paths;
 mod service;
 
+pub use finder::BlockPathFinder;
 pub use flow::find_flow;
 pub use package::find_package_file;
-pub use path_finder::BlockPathFinder;
 pub use search_paths::{get_block_value_type, BlockValueType};
